+++ source/Examples/Resources/CMDistributionGroup/CMDistributionGroup_Absent.ps1
@@ -1,0 +1,18 @@
+<#
+    .SYNOPSIS
+        A DSC configuration script to remove a distribution point group from Configuration Manager.
+#>
+Configuration Example
+{
+    Import-DscResource -ModuleName ConfigMgrCBDsc
+
+    Node localhost
+    {
+        CMDistributionGroup ExampleSettings
+        {
+            SiteCode          = 'Lab'
+            DistributionGroup = 'DistroGroup1'
+            Ensure            = 'Absent'
+        }
+    }
+}